Code refactoring for blog screen
