--- lib/blog/blogScreen.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'blog.dart';
-
-class BlogScreen extends StatelessWidget {
-  const BlogScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(body: BlogCard());
-  }
-}
